--- tests/Rector/Dynamic/FluentReplaceRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Dynamic\FluentReplaceRector: ~
