+++ tests/Rector/MethodBody/FluentReplaceRector/Correct/correct.php.inc
