+++ tests/Rector/MethodBody/FluentReplaceRector/Wrong/wrong.php.inc
