+++ tests/Rector/MethodBody/FluentReplaceRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\Rector\MethodBody\FluentReplaceRector: ~
